Updating and fixing some minor changes about message logging and exceptions checking in the CDS and the consumer and producer testers. Also, a docker-compose for testing purpose has been created.
