+++ docker/crypto-detector/crypto-consumer/requirements.txt
