+++ docker/crypto-detector/crypto-producer/requirements.txt
